backend/requirements.txt
# ECE Capstone
Django>=5.0
python-dotenv>=1.0.0
dj-database-url>=2.0.0
psycopg2-binary>=2.9.0
tensorflow>=2.15.0
tensorflow_hub>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
pytest>=7.0.0
tensorflow>=2.15.0
tensorflow_hub>=0.15.0  # MoveNet Thunder from TF Hub
numpy>=1.24.0
pandas>=2.0.0           # CSV loading, DataFrames
opencv-python>=4.8.0    # video/image loading
scikit-learn>=1.3.0     # data prep, train/test split, metrics
matplotlib>=3.8.0       # plotting training history
pytest>=7.0.0           # tests
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0
celery[redis]>=5.3.0
redis>=5.0.0
